Mostly PMT parser changes to match the VC-1/DTS changes on the C# side.

+++ DirectShowFilters/TsMuxer/source/IMuxInputPin.h
@@ -26,7 +26,7 @@ class IMuxInputPin
 {
   public:
     virtual byte GetId() = 0;
-    virtual int GetStreamType() = 0;
+    virtual byte GetStreamType() = 0;
     virtual DWORD GetReceiveTickCount() = 0;
     virtual HRESULT StartDumping(wchar_t* fileName) = 0;
     virtual HRESULT StopDumping() = 0;
